refactor: move Snake files under game/snake folder and fix all references
