+++ rn/src/game/snake/snake.ts
