+++ rn/src/game/snake/world.ts
